Add short pauses in the first test of each suite of RF Add expected status for Get on session for RF. an error otherwise Change order of tests in run suite Change Should Not Export Metrics More Then keyword
Short pauses in the first test of each suite allowed to immediately connect to the docker container instead of
the confusing connection attempts.

As for the order of tests in the run suite - theoretically it doesn't matter. In practice despite the fact that
previous workload finished successfully it still manages to mess up starting a new workload. Not seen elsewhere.

Should Not Export Metrics More Then keyword now waits for N seconds. takes a snapshot then waits 12 sec and checks that metrics file isn't changed.
This now prooves that no more metrics are gathered.